allow clippy because of macos

+++ agent/provider/src/hardware.rs
@@ -499,7 +499,8 @@ fn partition_space<P: AsRef<Path>>(path: P) -> Result<u64, Error> {
         use nix::sys::statvfs::statvfs;
         let stat =
             statvfs(path.as_os_str()).map_err(|e| sys_info::Error::General(e.to_string()))?;
-        Ok(stat.blocks_available() * stat.fragment_size())
+        #[allow(clippy::unnecessary_cast)]
+        Ok(stat.blocks_available() as u64 * stat.fragment_size())
     }
 }
